api_server: moved actions related structures
DeviceType and InstanceDeviceDetachAction are both structures
used for requests on /actions. These are now moved to the actions
module.
