+++ api_server/src/request/actions.rs
@@ -6,7 +6,7 @@ use futures::sync::oneshot;
 use hyper::Method;
 use serde_json::Value;
 
-use request::{InstanceDeviceDetachAction, IntoParsedRequest, ParsedRequest, SyncRequest};
+use request::{IntoParsedRequest, ParsedRequest, SyncRequest};
 
 // The names of the members from this enum must precisely correspond (as a string) to the possible
 // values of "action_type" from the json request body. This is useful to get a strongly typed
@@ -15,6 +15,20 @@ use request::{InstanceDeviceDetachAction, IntoParsedRequest, ParsedRequest, Sync
 pub enum ActionType {
     BlockDeviceRescan,
     InstanceStart,
+}
+
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+pub enum DeviceType {
+    Drive,
+}
+
+// Represents the associated json block from the sync request body.
+#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+#[serde(deny_unknown_fields)]
+pub struct InstanceDeviceDetachAction {
+    pub device_type: DeviceType,
+    pub device_resource_id: String,
+    pub force: bool,
 }
 
 // The model of the json body from a sync request. We use Serde to transform each associated
