+++ api_server/src/request/mod.rs
@@ -91,20 +91,6 @@ pub enum OkStatus {
     Updated,
 }
 
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub enum DeviceType {
-    Drive,
-}
-
-// Represents the associated json block from the sync request body.
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-#[serde(deny_unknown_fields)]
-pub struct InstanceDeviceDetachAction {
-    pub device_type: DeviceType,
-    pub device_resource_id: String,
-    pub force: bool,
-}
-
 impl GenerateResponse for OkStatus {
     fn generate_response(&self) -> hyper::Response {
         use self::OkStatus::*;
